+++ CHANGELOG.md
@@ -6,9 +6,13 @@
 
 组件支持watch方法，监听属性变化
 
+### 0.1.6（2017.2.24）
+
+去除widget.js的require，使用时可以直接copy这个文件到项目里。jsmin压缩，修改readme
+
 ### 0.1.5（2017.2.23）
 
-组件嵌套引用（a引用b，a引用c，b引用c）时，可以通过foo().addTo／foo().install避免冲突；
+组件嵌套引用（a引用b，a引用c，b引用c）时，可以通过foo().addTo／foo().install避免冲突
 
 ### 0.1.4-0.1.1